report_updater/requirements.txt
python-docx==1.1.2
openpyxl==3.1.5
PyYAML==6.0.2
streamlit==1.41.1
pytest==8.4.0
